set full article in new page and routing
