+++ SkyGazer/package.json
@@ -19,6 +19,7 @@
     "postcss": "^8.4.32",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.21.0",
     "tailwindcss": "^3.3.6"
   },
   "devDependencies": {
